+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.54.0] - 2026-06-17
+
+### Added
+
+- `tilebox-datasets` and `tilebox-workflows`: Added optional `transport="http1"` client configuration to use a
+  fallback protocol over HTTP/1.1 for networks that do not handle gRPC over HTTP/2 correctly.
+- `tilebox-worklows`: Added `runner.connect_to` as updated API for the direct runner mode.
+
 ## [0.53.0] - 2026-06-05
 
 ### Added
@@ -375,7 +383,8 @@ the first client that does not cache data (since it's already on the local file 
 - Released under the [MIT](https://opensource.org/license/mit) license.
 - Released packages: `tilebox-datasets`, `tilebox-workflows`, `tilebox-storage`, `tilebox-grpc`
 
-[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.53.0...HEAD
+[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.54.0...HEAD
+[0.54.0]: https://github.com/tilebox/tilebox-python/compare/v0.53.0...v0.54.0
 [0.53.0]: https://github.com/tilebox/tilebox-python/compare/v0.52.0...v0.53.0
 [0.52.0]: https://github.com/tilebox/tilebox-python/compare/v0.51.0...v0.52.0
 [0.51.0]: https://github.com/tilebox/tilebox-python/compare/v0.50.1...v0.51.0